validate not null input parameters on business controller

+++ src/main/java/com/enrique/reservatusalaback/controller/BusinessController.java
@@ -27,6 +27,9 @@ public class BusinessController {
 
     @PostMapping
     public ResponseEntity<?> add(@RequestBody Business business) {
+        if (business == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_OBJECT, HttpStatus.BAD_REQUEST);
+        }
         Business addedBusiness = businessService.add(business);
         if (addedBusiness == null) {
             return new ResponseEntity<>(ResponseCode.ALREADY_EXISTENT_USER, HttpStatus.CONFLICT);
@@ -41,6 +44,9 @@ public class BusinessController {
 
     @GetMapping("/findById")
     public ResponseEntity<?> findById(@RequestParam Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_ID, HttpStatus.BAD_REQUEST);
+        }
         Business business = businessService.findById(id);
         if (business == null){
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
@@ -50,6 +56,9 @@ public class BusinessController {
 
     @PutMapping
     public ResponseEntity<?> update(@RequestBody Business business) {
+        if (business == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_OBJECT, HttpStatus.BAD_REQUEST);
+        }
         Business updatedBusiness = businessService.update(business);
         if (updatedBusiness == null) {
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
@@ -59,6 +68,9 @@ public class BusinessController {
 
     @DeleteMapping
     public ResponseEntity<ResponseCode> deleteById(@RequestParam Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_ID, HttpStatus.BAD_REQUEST);
+        }
         if (businessService.deleteById(id) > 0) {
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
         } else {
@@ -68,6 +80,12 @@ public class BusinessController {
 
     @PostMapping("/addRoom")
     public ResponseEntity<?> addRoom(@RequestParam Long id, @RequestBody Room room) {
+        if (id == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_ID, HttpStatus.BAD_REQUEST);
+        }
+        if (room == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_OBJECT, HttpStatus.BAD_REQUEST);
+        }
         Room addedRoom = businessService.addRoom(id, room);
         if (addedRoom == null) {
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
@@ -77,6 +95,12 @@ public class BusinessController {
 
     @PostMapping("/addOperation")
     public ResponseEntity<?> addOperation(@RequestParam Long id, @RequestBody Operation operation) {
+        if (id == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_ID, HttpStatus.BAD_REQUEST);
+        }
+        if (operation == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_OBJECT, HttpStatus.BAD_REQUEST);
+        }
         Operation addedOperation = businessService.addOperation(id, operation);
         if (addedOperation == null) {
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
